add support for lockfiles on checkout
When checking out repositories, check if a file <filename>.lock.<ext>
exists for the first filename on the kas CLI. In case this file exists
and the --update option is not specified, automatically append this
file to the kas CLI before performing any other kas operations.

When --update is specified, the lockfile is ignored.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>