add support to call kas from different cwd
This patch adds a new parameter '--directory DIR' which specifies in which
directory kas should carry out all operations. The behavior is hereby
similar to how other tools like tar behave when relocating. By that, all
relative paths are considered relative to the DIR instead of the cwd.
The root repo dir is also set to DIR, in case no kas file is provided on
the kas cmdline. In case the parameter is not specified, all paths are
relative to the cwd, which is identical to the behavior prior to this patch.

This is needed as a preparation to solve include path errors of the menu
plugin. There, an explicit anchor of the root repository is needed as
the generated kas file is written to the KAS_WORK_DIR.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>