menu plugin: fix combination with KAS_WORK_DIR
This patch fixes the combination of KAS_WORK_DIR and kas menu.
The default location of the Kconfig file is now relative to
the directory specified via '--directory DIR', or to the cwd
when not specified. In addition, the information about the DIR
is passed to the next kas stages (e.g., build).

Reported-by: Peter Hoyes <Peter.Hoyes@arm.com>
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>