kas-container: add support to call kas from different cwd
This patch adds support for the -C, --directory parameter in
kas-container. By that, the location of the first file on the kas
cmdline is no longer considered to be the root repo dir. Instead, the
cwd or the provided switch-to directory is considered the root repo dir.
This is in sync with the behavior in kas itself.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>