kas-container: add support to call kas from different cwd
This patch adds support for the -C, --directory DIR parameter in
kas-container. By that, all relative path are considered relative to DIR
instead of the CWD. This is technically achieved by cd-ing into DIR
before any further operation. The logic to locate the root repo
directory by querying the CVS of the first kas file for its base is not
affected. Further, this parameter is not passed to kas inside the
container, as the repo mount paths inside the container are not affected.
This is in sync with the behavior in kas itself.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>