kas-container: track REPO dir when using menu cmd
When using the kas menu command, we need to track the repository
location outside of the kas container as well. This is required to
prepare the environment for a later kas-container build, as the repo
path inside the container is different from the one outside. For that,
the location of KAS_REPO_DIR is written to a .kas-container.conf file
next to the .config.yaml file. When running kas-container build, we
check if this config file is found in the KAS_WORK_DIR and source it.
This sets the KAS_REPO_DIR variable to the same value that was used when
invoking kas-container menu.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>