fix: correctly handle upstream ff merges on fetch
In case both branch and commit are specified, kas checks if the commit
is contained in the branch. Further, only when a ref was not present in
the local clone, a fetch from remote was executed. This logic however
breaks on upstream fast-forward merges, as these do not create a new
commit id, but instead just move the branch head.

This commit adds a check if we also have the requested commit sha on the
requested branch locally and if not continues to the fetching step
(instead of the short circuit).

Fixes: #122
Reported-by: Jan Kircher <jan.kircher@leica-microsystems.com>
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>