graceful cancellation of external commands
Instead of just propagating the sigint / sigterm to the subprocess
executor, we use the asyncio infrastructure to send a cancellation event
to all tasks. By using asyncio.gather instead of .wait we ensure all
downstream tasks are cancelled on the first failure (or cancellation
event), making sure to not leave any orphaned tasks behind.
To make sure the cancellation signal (e.g. SIGINT) arrives in the
correct order, we spawn the subprocesses in their own process group.
This leads to the following signal chain:

1. user presses CTRL^C -> SIGINT sent to Python process
2. signal handler cancels all tasks
3. asyncio raises a CancelledError inside coroutines, which results in
   a SIGINT being sent to the external process
4. execution continues once all external processes terminated

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>