fix: find missing repos by id instead of name
In the repo dict, repositories are stored by id (not by name). Also all
repository lookup functions perform the lookup by id instead of name
(despite always calling this parameter "name"). This makes sense, as the
id has to be unique (enforced by the schema), while the name could be
duplicated.

When looking up repos to check if they are already fetched, these were
stored by name instead of id. By that, repos that had a name name
could not be looked up. We fix this by storing the key along with the
repo in the lookup list and always compare by key.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>