tests: make create ref repo test git only
In fact only git was tested before, but the mercurial repo was
referenced in our test file. We now switch over to another test file
which tests this logic even better as it contains the kas repo twice
(while the reference must only be there once).

This is a preparation to rework our mercurial tests.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>