kas-container: Warn upon duplicated user ids
Currently Dockerfile creates the builder user and later
upon container execution the user id of this user is changed
to the user id passed by kas-container.
As soon as the passed user id conflicts with another user
within the container we are in trouble as two users with
different name and home directory use the same uid.

Print a warning in such a situation to make the user aware
of possible problems.

Signed-off-by: Richard Weinberger <richard@nod.at>