kas-container: Preserve supplementary groups
Currently kas-container passes user and group id to container-entrypoint
which uses gosu to switch to this uid/gid pair.
But supplementary groups are omitted. This causes access problems on
system with non-trivial permissions such as ACLs
In my case access to DL_DIR fails since access is granted by ACL which
involve supplementary groups.

Since gosu is already aware of supplementary groups we can pass the whole
group list to container-entrypoint and assign them to the builder user.
For each supplementary group id a group with name "g<GID>" is created.

Signed-off-by: Richard Weinberger <richard@nod.at>