forcefully terminate on second CTRL-C
The first CTRL-C cancels all tasks. If the tasks are external commands
issued via run_cmd_async, these get a SIGINT and we wait until they
terminate (which can be indefinitely long). If the user presses CTRL-C
a second time, a CancelledError exception is thrown inside the other
(waiting) exception, leading to chaotic shutdown of the event loop and
internal exceptions thrown in the destructors (which cannot be caught
and spam the console).

We fix this by replacing the signal handler on the first CTRL-C with one
that immediately terminates kas without any further interaction with the
event loop. While this leads to an unclean shutdown, it still is the best
we can do as it just hands over the proper termination of the processes
to the OS.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>