kas-container: support git worktrees
Git support sharing the .git directory across multiple worktrees each having
its own HEAD, index, rebase-todo and so on. The worktree directory (created
with `git worktree`) contains no directory .git, but a file .git with a
reference to the main git directory. Therefore, this directory must also be
mounted in the container to make git operable.

[Felix]
Git uses an absolute path to anchor the worktrees to the common
repository. As we do not want to recreate the full host path inside the
container, we mount the common repo as /repo-common and patch the .git
file of the worktree to point to that instead of the path on the host.
This is implemented using a temporary .git file with the patched path.
By that, the worktree on the host is kept unchanged.

Closes: #107
Signed-off-by: Jörg Sommer <joerg.sommer@navimatix.de>
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>