lock plugin: update corresponding lockfiles
As we now track from which lockfile a repos lock originates, we also
want to update it there. This is a two-staged process:

1. update all locks of "locked" repos
2. add locks for not-yet-locked repos to top lockfile

In step 2 we further need to handle the case that the default lockfile
does not yet exist, and the case that it already exists. In case it does
not exist, it will be created. Otherwise it will be updated.

Co-developed-by: Frieder Schrempf <frieder.schrempf@kontron.de>
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>