attestation: only include config if not tracked
If the configuration file (kas project config) is tracked, we do not
need to include it in the attestation data. By that, we reduce the size
of the attestation and also avoid a potential leak of information in
case the receiver of the attestation data should not have access to the
repo (and project config file).

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>