extend kas schema to provide repo signing information
As of now kas does not provide any mechanisms to verify the integrity of the
git repos (beyond the git commit ids). There had been attempts in the past to
add this functionality by checksums over the repo contents, but these were
not accepted due to the complexity of the implementation. Also these do not
help against updating the repo to a different commit id.

This commit prepares the kas schema to specify cryptographic signing
information (signers, allowed_signers) that later can be used to validate
the signature of a given repository.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>