load external lockfiles also on checkout --update
When running the checkout plugin (or any derived plugin) with --update,
lockfiles were not considered. This logic ensures that the repos can be
updated and later-on be pinned to the new revisions (by updating the
lockfiles). However, any external lockfile (from an external repo)
should always be loaded, as these are not under our control and we
cannot update them. By that, it also makes no sense to move the
externally-pinned repos forward as the next kas invocation without
--update will bring them back to the pinned state.

To fix this, we now always load the lockfiles of external repos (by
that, also on --update).

Note: This introduces a slight behavior change, but only on executions
with --update.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>