kas-container: add prune option to cleanup all created files
The cleanup commands only remove the files created by bitbake, but not
the ones created by kas. This is a problem in rootless execution, as
these files belong to a sub uid/gid, as they were created by the
non-root user inside the container. By that, they cannot be removed from
the host (by a non-privileged user). To fix this, we provide a new
kas-container command "prune" that removes all data that is under kas
control. In case of rootless, the rm is performed inside the container
as there the uids are matching.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>