add family of commands for cleaning
Previously we had the clean / cleansstate / cleanall commands only
implemented for the kas-container. This proved not sufficient under some
scenarios (like rootless), as the cleaning needs to happen from inside
the container. Anyways, it was an inconsitency between kas-container and
kas.

We now add support for these commands in kas as well.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>