add kas purge command to remove all managed data
The kas purge command is similar to kas cleanall, but also removes all
repos managed by kas (including the corresponding repos in
KAS_REPO_REF_DIR). Once executed, no kas or bitbake managed data remains
in any of the provided directories (whereby managed is refering to the
current configuration and environment settings).

The user specified dirs are not removed, as these are managed by the
calling user and could be mountpoints.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>