docs: precisely describe how config files are merged
There has been some ambiguity around how exactly the checkout and
handling of external lockfiles works. This became worse with the
introduction of lockfiles as it was not specified when exactly lock
files are loaded.

We now document the current behavior of kas and also fix a slight
documentation bug: Technically, the lock files are loaded prior to the
file they correspond to (from the perspective of the merge logic). This
is in-line with the handling of include files which are also processed
depth first (hereby the lock file is similar to an include file).

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>