GitLab CI: add git safe dir for project root
The GitLab CI runner mounts the $CI_PROJECT_DIR git project root from
the outside but does not align the ownership of that directory with the
user of the docker container. By that, git does not allow to perform any
operation on that repository.

As this is a well-known case, we mark that directory as a safe dir if
needed and running in GitLab CI.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>