run_cmd: return triple on error when capturing stderr
In case we run run_cmd with capture_stderr=True and run into an error
like FileNotFound or Permission, we also need to return a triple instead
of a tuple. Otherwise the return type is not the same depending on if an
error occured or not. This further fixes the issue that _filter_stderr
is called without a fourth parameter (stderr) in run_cmd_async.

These fixes do not have any user-facing impact.

Fixes: e8e5ea54a ("run_cmd: add support to capture stderr as well")
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>