unify processing of env vars that point to a path
The handling and checking of environment variables that point to a path
(e.g. NETRC_FILE) was done quite inconsitently across kas. By that, on
some variables that point to an non existing path, internal exceptions
were raised.

We unify this by adding a new EnvSetButNotFoundError error that is
reported by kas with a meaningful error message along with a function to
check the error conditions. This also simplifies the code.

Reported-by: Christoph Steiger <christoph.steiger@siemens.com>
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>