entrypoint: copy userdata and chown on rootless
User provided credential files are mounted into /var/kas/userdata.
Hereby kas-container sets the env vars accordingly and kas just reads
them from there. This breaks on docker rootless, as the builder user
might not be able to read the host-mounted files, as they are mounted as
root into the container.

We fix this by moving the mountpoint /var/kas/userdata to userdata.orig,
then copy the contents back to userdata and chown them there. By that,
the variables that point to the credential files do not need to be
changed.

Reported-by: Aidan Moss <aidanmoss15@gmail.com>
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>