do not change current working directory on setup_dir
The SetupDir command takes care of creating the kas build directory and
also changes the current working directory (despite not mentioned in the
docstring). This is is problematic as there is no way to later tell
which directory kas was initially started in.

When setting up the initial config (happens before running the macros),
we resolve the kas files relative to the working directory - which is
expected to be the one kas was invoked in. However, when creating a
second set of config files after the first has been created and SetupDir
was executed, this fails as we try to resolve relative to KAS_WORK_DIR.
In current kas this behavior is side-effect free, as it only operates on
a single config. However, with upcoming plugins like diff not globally
changing the cwd becomes mandatory.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>