test: use parameterized dirs for marked tests
We currently lack test coverage of the scenarios where KAS_WORK_DIR or
KAS_BULID_DIR is set. To improve this, we parameterize the monkeykas
fixture to test the variants "no env", "KAS_WORK_DIR is set" and both
"KAS_(WORK/BUILD)_DIR is set".

As this results in a multiplication of the executed tests, we only
enable this fanout for tests that are marked with the dirsfromenv
marker.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>