diff: improve formatting of dict changes
Previously, changed keys values were reported as removed / added, but
the diff was not performed within the value itself. Further, keys that
were added were also reported as removed with value 'None' before (and
vice versa for removed keys).

We change that in the following way:

- added / removed keys are only reported once
- changes of a key value are diffed by value (using difflib)

We further document, that the text output is not stable. If stable
output is needed, the json format shall be used.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>