warn if kas-container version does not match kas version
As a preparation to introduce a stable interface between kas and
kas-container, we add a check in kas that warns the user if the versions
do not match. Using kas with a non-matching kas-container version was
never supported, but as the interface between both was rather stable,
only little issues were brought up. By that, we use a warning instead of
an error if the versions do not match.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>