deprecate various magic values to disable a layer
Instead of allowing a variety of strings to disable a layer, we stick to
a single boolean value (false). By that, we only provide a single way to
disable a layer, which both helps the IDEs, as well as avoids ambiguity.

To give downstream users some time to migrate, we still allow the old
values, but file a deprecation warning if used.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>