clean: add option to recover from disk-full situation
When running kas clean with a provided config file, kas needs to setup
the environment to process that config. This requires some temporary
space, which is not available on a full disk. By that, the clean fails
and makes it impossible to recover from that situation (at least with
kas mechanisms).

We now add an option --recover to clean (and alike), that avoids the
Macro execution and by that can be run on a 100% full disk.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>