kas-container: add experimental support for run0
ISAR builds need to be executed in privileged mode. Currently,
kas-container uses sudo to start the container, but this is breaks
in environments that set the "no new privileges" flag (like the
VSCode shell). This can be solved by using run0 (which is a wrapper
around systemd-run) to let the init system directly start the
privileged workload (with PAM authentication and alike).

We now add experimental support for run0 to kas-container for targets
that conceptionally support this (like the podman / docker execution).
We further add this as a fallback in case sudo is not installed.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>