kas(isar): add support to clean with run0
In isar mode, a privileged executor is needed to remove the tmpdir(s).
As of now, isar requires sudo, but we already prepare the code to be
run0 compatible as well. By that, we can re-use the KAS_SUDO_CMD
environment variable.

Note, that the variable is not forwarded into the container, as the
tooling inside the container is independent from the outside (e.g.
outside you might want to use run0, while inside you have to use
whatever is provided by the container).

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>