docs: add oe-core nodistro example
OpenEmbedded starts to split the poky repository into individual
components (bitbake and core). By that, bitbake does not come along with
the layers, but as individual repository. We already support this case,
however writing a configuration for that case might be non trivial for
newcomers. To help them, we add an example.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>