fix(repos): raise on fetching errors
We currently only warn in case the fetching fails. However, we must not
continue as otherwise kas assumes the repo is up-to-date and passes
incorrect data to downstream consumers (like dump, lock). Before the
introduction of the dump / lock command, this was merely an annoyance as
the checkout did not match the upstream state. With the consumers, it is
a bug.

While changing this, we also remove all error-code returns in that
function which never have been checked for anyways.

Fixes: daf0abab5 ("Initial public release")
Closes: #168
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>